requirements.txt

# Core dependencies
pyyaml>=6.0
torch
transformers
datasets
python-dotenv
# Optional: For Gemini API support
google-generativeai>=0.3.0

# Optional: For local LLM support (choose one or more)
# vllm>=0.2.0  # For vLLM inference
# openai>=1.0.0  # For OpenAI-compatible API calls
# transformers>=4.30.0  # For direct model loading
# torch>=2.0.0  # Required for transformers

# Development dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0